+++ RedUtils/Actions/DoubleJumpShot.cs
@@ -1,12 +1,11 @@
 ﻿using System;
 using System.Timers;
 using RedUtils.Math;
-using RedUtils.Objects;
 
 namespace RedUtils
 {
-    /// <summary>A double jump shot action, where the car double jumps into the ball</summary>
-    public class DoubleJumpShot : Shot
+	/// <summary>A double jump shot action, where the car double jumps into the ball</summary>
+	public class DoubleJumpShot : Shot
 	{
 		/// <summary>Whether or not this shot has finished</summary>
 		public override bool Finished { get; internal set; }
@@ -26,7 +25,7 @@ namespace RedUtils
 		public Arrive ArriveAction { get; internal set; }
 
 		/// <summary>The amount of boost we have when starting this action</summary>
-		private readonly int _startBoostAmount = 0;
+		private readonly float _startBoostAmount = 0;
 		/// <summary>The length between updates to the target location, and shot direction</summary>
 		private readonly float _updateInterval = 0.2f;
 		/// <summary>Keeps track of the time since the last update</summary>
